% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_M3_quarterly}
\alias{load_M3_quarterly}
\title{load M3 quarterly time series}
\usage{
load_M3_quarterly(file = "G:/Mi unidad/data/m3_quarterly_dataset.tsf")
}
\arguments{
\item{file}{The tsf file with the data}
}
\value{
A list with the time series in ts format
}
\description{
load M3 quarterly time series
}
