% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_M3_yearly}
\alias{load_M3_yearly}
\title{load M3 yearly time series}
\usage{
load_M3_yearly(file = "G:/Mi unidad/data/m3_yearly_dataset.tsf")
}
\arguments{
\item{file}{The tsf file with the data}
}
\value{
A list with the time series in ts format
}
\description{
load M3 yearly time series
}
