% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_M4_yearly}
\alias{load_M4_yearly}
\title{load M4 yearly time series}
\usage{
load_M4_yearly(file = "G:/Mi unidad/data/m4_yearly_dataset.tsf")
}
\arguments{
\item{file}{The tsf file with the data}
}
\value{
A list with the time series in ts format
}
\description{
load M4 yearly time series
}
